# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Booking API documentation",
    "description": "L5 Swagger OpenApi description",
    "contact": {
      "email": "khujand.nanocoding@gmail.com"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001/api",
      "description": "Local API Server"
    }
  ],
  "paths": {
    "/booking": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Display all bookings",
        "operationId": "bookingAll",
        "parameters": [
          {
            "name": "table_id",
            "in": "query",
            "description": "The booking`s table id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "user_id",
            "in": "query",
            "description": "The booking`s user id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "in": "query",
            "description": "The booking`s date",
            "schema": {
              "type": "date"
            }
          },
          {
            "name": "time_start",
            "in": "query",
            "description": "The booking`s time start",
            "schema": {
              "type": "time"
            }
          },
          {
            "name": "time_end",
            "in": "query",
            "description": "The booking`s time end",
            "schema": {
              "type": "time"
            }
          },
          {
            "name": "phone",
            "in": "query",
            "description": "The booking`s phone",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "description",
            "in": "query",
            "description": "The booking`s description",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "The booking`s statuses",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BookingModel"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Create a booking",
        "operationId": "bookingPost",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/booking/{id}": {
      "get": {
        "tags": [
          "Booking"
        ],
        "summary": "Show booking by id",
        "operationId": "showBookingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The booking`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Booking"
        ],
        "summary": "Update booking by id",
        "operationId": "updateBookingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The booking`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BookingModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Booking"
        ],
        "summary": "Delete booking by id",
        "operationId": "deleteBookingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The booking`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/booking/toggle/{id}": {
      "post": {
        "tags": [
          "Booking"
        ],
        "summary": "Toggle status booking by id",
        "operationId": "toggleStatusBookingById",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "The booking`s id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "404": {
            "description": "Not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Create a user",
        "operationId": "userSignup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterStoreModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedModel"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/signin": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login to site",
        "operationId": "userSignin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoggedModel"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BookingModel": {
        "title": "Booking model",
        "description": "Booking model",
        "properties": {
          "id": {
            "title": "Id",
            "description": "Booking`s id",
            "format": "string"
          },
          "table_id": {
            "title": "Table id",
            "description": "Booking`s table id",
            "format": "integer"
          },
          "date": {
            "title": "Date",
            "description": "Booking`s date",
            "format": "date"
          },
          "time_start": {
            "title": "Time start",
            "description": "Booking`s time start",
            "format": "time"
          },
          "time_end": {
            "title": "Time end",
            "description": "Booking`s time end",
            "format": "time"
          },
          "phone": {
            "title": "Phone",
            "description": "Company`s phone",
            "format": "string"
          },
          "description": {
            "title": "description",
            "description": "Booking`s description",
            "format": "string"
          },
          "user_id": {
            "title": "User id",
            "description": "Booking`s user id",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Booking`s status",
            "format": "string"
          },
          "created_at": {
            "title": "Created at",
            "description": "Booking`s created at",
            "format": "date-time"
          },
          "updated_at": {
            "title": "Updated at",
            "description": "Booking`s updated at",
            "format": "date-time"
          }
        },
        "type": "object"
      },
      "BookingStoreModel": {
        "title": "Booking store model",
        "description": "Booking store model",
        "properties": {
          "table_id": {
            "title": "Table id",
            "description": "Booking`s table id",
            "format": "string"
          },
          "date": {
            "title": "Date",
            "description": "Booking`s date",
            "format": "date"
          },
          "time_start": {
            "title": "Time start",
            "description": "Booking`s time start",
            "format": "time"
          },
          "time_end": {
            "title": "Time end",
            "description": "Booking`s time end",
            "format": "time"
          },
          "phone": {
            "title": "Phone",
            "description": "Company`s phone",
            "format": "string"
          },
          "description": {
            "title": "description",
            "description": "Booking`s description",
            "format": "string"
          },
          "user_id": {
            "title": "User id",
            "description": "Booking`s user id",
            "format": "string"
          }
        },
        "type": "object"
      },
...
      "LoggedModel": {
        "title": "Logged model",
        "description": "Logged model",
        "properties": {
          "user": {
            "title": "user",
            "description": "Users`s data",
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "email"
              },
              "phone": {
                "type": "string"
              },
              "image": {
                "type": "string"
              },
              "role": {
                "type": "string"
              },
              "company_id": {
                "type": "string"
              },
              "status": {
                "type": "string"
              },
              "updated_at": {
                "type": "string"
              },
              "created_at": {
                "type": "string"
              }
            },
            "type": "object",
            "format": "object"
          },
          "token": {
            "title": "Access token",
            "description": "Users`s access token",
            "format": "string"
          }
        },
        "type": "object"
      },
      "LoginModel": {
        "title": "Login model",
        "description": "Login model",
        "properties": {
          "email": {
            "title": "Email",
            "description": "Users`s email",
            "format": "email"
          },
          "password": {
            "title": "Password",
            "description": "Users`s password",
            "format": "string"
          }
        },
        "type": "object"
      },
...
      "RegisterStoreModel": {
        "title": "Register store model",
        "description": "Register store model",
        "properties": {
          "name": {
            "title": "Name",
            "description": "Users`s name",
            "format": "string"
          },
          "email": {
            "title": "Email",
            "description": "Users`s email",
            "format": "email"
          },
          "password": {
            "title": "Password",
            "description": "Users`s password",
            "format": "string"
          },
          "role": {
            "title": "Role",
            "description": "Users`s role",
            "format": "string",
            "default": "user",
            "enum": [
              "admin",
              "user"
            ]
          },
          "phone": {
            "title": "Phone",
            "description": "Users`s phone",
            "format": "string",
            "example": "+992927777777"
          },
          "image": {
            "title": "Image",
            "description": "Users`s image",
            "format": "string",
            "example": "no-image.png"
          },
          "company_id": {
            "title": "Company id",
            "description": "Users`s company id",
            "format": "string"
          },
          "status": {
            "title": "Status",
            "description": "Users`s status",
            "format": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}